requirements.txt
flask
flask-sqlalchemy
flask-migrate

Pillow